Add start to ffmpeg_translate script
It's a start. It currently accepts rtmp streams for input and
output, and alsa devices for translator input and monitor.

It will mix the normalized translator input audio into the video
audio while ducking the video audio. The video audio will be played
in the translator monitor device.

Known issue: If input stream is restarted, timestamps
are out of sync. This needs to be fixed.